test: create org.dhis.feebdack.Assertions
